* uring: (41 commits)
  fuse-io-uring: Fix possibly uninitialized values
  fuse-io-uring: Add fuse_uring_set_payload_allocator()
  Avoid nested function declarations in helper functions
  mount_bsd: Remove unused headers
  mount_bsd: Show errors when syscalls failed
  mount_bsd: Proper exit calls
  FreeBSD: Remove useless options
  mount_bsd: Fix usage of libfuse_strtol
  tests: Re-enable mknod and mkfifo tests on FreeBSD
  tests: Disable tests with TMP_FILE on FreeBSD
  build(deps): bump github/codeql-action from 3.28.8 to 3.28.9
  Add comment for fuse_passthrough_open()
  ci-build test: Add a 32-bit compilation test
  tests: Skip the ioctl test for x86 cross compilation
  example/hello_ll_uds: Switch to %zu and avoid 32bit build warning
  fusermount: Exclude UFSD from whitelist for 32-bit builds
  Fix build of example/memfs_ll.cc on 32 bit architectures
  Add signify key for 3.18
  passthrough_hp: enable splice move by default
  Prepare 3.17.1 release
  ...

+++ example/memfs_ll.cc
@@ -150,7 +150,7 @@ class Inode {
 
 	void read_content(char *buf, size_t size, off_t offset) const
 	{
-		size_t bytes_to_read = std::min(size, content.size() - offset);
+		size_t bytes_to_read = std::min(size, content.size() - (size_t)offset);
 		std::copy(content.begin() + offset,
 			  content.begin() + offset + bytes_to_read, buf);
 	}
@@ -613,7 +613,7 @@ static void memfs_read(fuse_req_t req, fuse_ino_t ino, size_t size,
 	}
 
 	std::vector<char> content(
-		std::min(size, inode->content_size() - offset));
+		std::min(size, inode->content_size() - (size_t)offset));
 	inode->read_content(content.data(), content.size(), offset);
 
 	inode->unlock();
